+++ components/frontend/src/subject/SubjectTableRow.js
@@ -26,6 +26,9 @@ export function SubjectTableRow(
         report,
         reportDate,
         reports,
+        showIssueCreationDate,
+        showIssueSummary,
+        showIssueUpdateDate,
         subject_uuid,
         hiddenColumns,
         visibleDetailsTabs,
@@ -94,7 +97,15 @@ export function SubjectTableRow(
             {!hiddenColumns.includes("unit") && <Table.Cell style={style}>{unit}</Table.Cell>}
             {!hiddenColumns.includes("source") && <Table.Cell style={style}><MeasurementSources metric={metric} /></Table.Cell>}
             {!hiddenColumns.includes("comment") && <Table.Cell style={style}><div dangerouslySetInnerHTML={{ __html: metric.comment }} /></Table.Cell>}
-            {!hiddenColumns.includes("issues") && <Table.Cell style={style}><IssueStatus metric={metric} issueTracker={report.issue_tracker} /></Table.Cell>}
+            {!hiddenColumns.includes("issues") && <Table.Cell style={style}>
+                <IssueStatus
+                    metric={metric}
+                    issueTracker={report.issue_tracker}
+                    showIssueCreationDate={showIssueCreationDate}
+                    showIssueSummary={showIssueSummary}
+                    showIssueUpdateDate={showIssueUpdateDate}
+                />
+            </Table.Cell>}
             {!hiddenColumns.includes("tags") && <Table.Cell style={style}>{get_metric_tags(metric).map((tag) => <Tag key={tag} tag={tag} />)}</Table.Cell>}
         </TableRowWithDetails>
     )